Made some typo/nit corrections. Reordered some stuff

+++ overloads/vector.h
@@ -111,8 +111,8 @@ vector<T> exponent(const vector<T> &a) {
 // isEquals operator
 template<typename T>
 bool operator == (const vector<T> &a, const vector<T> &b) {
-    int n = a.size();
-    int m = b.size();
+    size_t n = a.size();
+    size_t m = b.size();
 
     if(n != m) {
         return false;
